% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{combine_collection}
\alias{combine_collection}
\title{Download page collect statistics}
\usage{
combine_collection(link, motorrad)
}
\arguments{
\item{link}{URL of the desired page}

\item{motorrad}{Character vector of length 1, containing the name of the
motorcycle}
}
\description{
Download page collect statistics
}

